Move code from user to team because both users is always used with team
