+++ maelstrom/gameplay/combat.py
@@ -7,12 +7,11 @@ from functools import reduce
 from maelstrom.campaign.level import Level
 from maelstrom.dataClasses.character import Character
 from maelstrom.dataClasses.activeAbilities import TargetOption
-from maelstrom.dataClasses.team import Team
+from maelstrom.dataClasses.team import Team, User
 from maelstrom.dataClasses.weather import WEATHERS, Weather
 from maelstrom.loaders.character_loader import EnemyLoader
 from maelstrom.ui import AbstractUserInterface, Choice, Screen
 from maelstrom.util.stringUtil import lengthOfLongest
-from maelstrom.util.user import User
 
 import random
